remove gravity dependency
code cleanup

+++ include/constants.h
@@ -6,14 +6,14 @@ bool initialized = false;
 struct constants
 {
     //tolerance variables
-    int air_humidity_min;
-    int air_humidity_max;
+    float air_humidity_min;
+    float air_humidity_max;
 
-    int air_temperature_min;
-    int air_temperature_max;
+    float air_temperature_min;
+    float air_temperature_max;
 
     float ec_reading_min;
-    int ec_reading_max;
+    float ec_reading_max;
 
     float ph_reading_min;
     float ph_reading_max;
